+++ scripts/sheets/template-components.js
@@ -61,9 +61,9 @@ export class TemplateComponents {
     editlocation,
     isOwner = false,
   ) {
-    const systemClass = game.system.id === "dnd5e" ? " dnd5e" : "";
+    const systemClass = game.system.id === "dnd5e" ? " dnd5e2" : "";
     const journalClass =
-      game.system.id === "dnd5e" ? "journal-entry-content" : "";
+      game.system.id === "dnd5e" ? "journal-entry-content dnd5e2 themed theme-light" : "";
 
     if (!isOwner)
       return `